backend/requirements.txt
torch>=1.9.0
nltk>=3.6.2
git+https://github.com/boudinfl/pke.git
spacy>=3.1.0
numpy>=1.20.0
flashtext>=2.7
transformers>=4.9.1
sense2vec>=2.0.0
sentence-transformers>=2.2.0
python-Levenshtein>=0.12.2
pdfplumber>=0.7.0
flask>=2.0.1
flask-cors>=3.0.10
python-dotenv>=0.19.0
requests>=2.25.0
